[NCL-1483] Add optional field with repositories
This optional field is added to the Product import process.

This allows products that need custom repositories (not defined in the
pom.xml) to be analyzed by allowing the user to specify the custom
repositories.

On submit, the additional field 'repositories' will be sent to the
configured Dependency Analysis server. That field will contain a list of
repositories to add.